futility: Fix HWID digest footer output
CL:4624937 contained a small error: the footer for print_hwid_digest()
gets printed in the parseable case when it was meant for the readable
case. This causes the readable output to be missing a line break and the
parseable output to include an extra empty `bios::GBB::` line.

The footer is only ever used to print a newline anyway, and therefore
pointless. This patch gets rid of it and fixes the bug.

BRANCH=none
BUG=none
TEST=make runtests

(cherry picked from commit 774c700fe1e5a8b1ac0f75a94d719b780af1b2eb)

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I8969c403faddbf97446fb33de9ba3c874b316336
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4691088
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-by: Saketh Pothireddy <spothire@google.com>
GitOrigin-RevId: 774c700fe1e5a8b1ac0f75a94d719b780af1b2eb
Cr-Build-Id: 8739997580243611025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739997580243611025
Copybot-Job-Name: android-main-fw-nissa-vboot_reference-copybot-downstream