futility: Add basic README.md
Add basic markdown syntax README.md and link with existing GBB
subcommand doc.

We already have the usage function help text (by `futility --help`), but
we would like to have another non-runtime document that can be used as a
reference in other documents. The usage context and requirements are
different, so a non-runtime README.md would be helpful.

We started with the README.md that was almost the futility usage
function. We removed the outdated information and add more structural
information and examples to be more informative.

BUG=b:298146304
TEST=pandoc README.md -o output.html
BRANCH=none

Change-Id: I5d6a8b146f1eaeb3f34f26f3e64e3396ff39da30
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>