futility: updater: Do not update RO when the AP RO is locked
For systems with SI_DESC that locked and disabled the write permission
fo CPU (AP), previously we will ignore updating SI_DESC and update the
rest of RO sections. However that will fail if the system also has
AP RO verification v2 (RO_GSCVD) because the hash calculation includes
SI_DESC.

As a result, we want to skip updating RO when the AP RO is locked. This
is currently implemented as a new default quirk 'check_locked_ap_ro'.

BUG=b:284913015
TEST=FEATURES=test emerge vboot_reference
     Get two Nissa firmware images as "locked.bin" and "unlocked.bin":
     futility update -i locked.bin --emu locked.bin --debug
     futility update -i unlocked.bin --emu unlocked.bin --debug
     The log shows correct FLMSTR1 values (and correct update modes).
BRANCH=None

Change-Id: Ieb76e1615635768f542defbd91afcd076e08f961
Signed-off-by: Hung-Te Lin <hungte@chromium.org>