futility: update: Refactor 'unlock ME' quirks
The quirk 'unlock_me_for_update' and 'unlock_wilco_me_for_update' are
doing the same thing with different patterns, but the implementation was
quite different. With the new `find_fmba`, we can unlock by changing
FLMSTRs in a unified approach. In the future we should also merge the
`unlock_flash_master` to follow the same implementation.

Also fixed a bug that the test data was using Link that the FLMSTRs are
actually in 0x60 instead of the hard-coded 0x80.

BUG=None
TEST=FEATURES=test emerge vboot_reference
BRANCH=None

Change-Id: I1d6a4cf6be2c5af9562914841c27f2b9a4c02adc
Signed-off-by: Hung-Te Lin <hungte@chromium.org>