futility: update: Refactor the 'unlock ME' quirk(s)
The quirk 'unlock_me_for_update' (and the deprecated quirk
'unlock_wilco_me_for_update') and unlock_flash_master are doing almost
the same things with different patterns, but the implementations were
quite different.

With the new `find_fmba`, we can unlock by changing FLMSTRs in a
unified approach. The first step is to change 'unlock_me_for_update'
(which works for glados=skylake and eve=kabylyake platforms) to
unlock_me_eve (because the support for glados is already deprecated),
and change the implementation to accept different parameters. In the
follow up commits we will also merge the unlock_flash_master (designed
for nissa) to the quirk system, using the same implementation.

Also fixed a bug that the test data was using Link that the FLMSTRs are
actually in 0x60 instead of the hard-coded 0x80.

BUG=None
TEST=FEATURES=test emerge vboot_reference
BRANCH=None

Change-Id: I1d6a4cf6be2c5af9562914841c27f2b9a4c02adc
Signed-off-by: Hung-Te Lin <hungte@chromium.org>