futility/updater: Fix EC software write protection logic
In CL:4301750, we attempted to reuse the function
is_write_protection_enabled() for both AP and EC write protection.
However, this was not correct because dut_get_property(DUT_PROP_WP_SW,
cfg) will always use the internal programmer and thus get the AP
software write protection instead of EC's. This caused the issue that we
have been falsely using the AP SWWP as EC SWWP since that CL.

If you are attempting to update the EC using futility with the following
settings: HWWP=1, AP SWWP=0, EC SWWP=1, and the EC_RO region in the
target image is different from the DUT's, futility will believe that it
can fully update the entire EC (RO+RW), but it actually can't. This will
cause flashrom to attempty to flash both RO+RW, notice that flashing RO
failed, verify the EC_RO region, and then error out.

To fix this, we need to re-implement the EC write protection logic to
get the real write protection value.

BUG=b:287933146
BRANCH=none
TEST=cq

Change-Id: Idd86e7cae3483f3864ef54e0116f91604133ecc3
Signed-off-by: roccochen@chromium.com <roccochen@chromium.org>