strip_boot_from_image.sh: sfill fast
We're writing two passes here, one with random data and the next with
zeros. Since we're not actually trying to secure delete anything we can
abide by not using urandom or sync writes. This speeds up the operation
from a reported 10MB/s to 100+MB/s based on other's reported findings.

BUG=b:295358776
TEST=local signing appears ok && args succeed

Change-Id: Ib68fd60dee7c40f10533a24d955ba019007f56df