firmware/2lib: Montgomery mult. using modulus positive inverse
This commit implement the Algorithm 2 as described in Montgomery
Multiplication Using Vector Instructions from August 20, 2013
(cf. https://eprint.iacr.org/2013/519.pdf). This variation of the
Montgomery multiplication algorithm decouple some arithmetic steps
which as a result be performed in parallel to improve performances.

This new algorithm relies on different pre-computed data. Indeed, it
uses the modulus positive inverse (n0pinv = 1 / N[0] mod 2^32) instead
of the modulus negative inverse. As a result, the public key data
structure needs to be updated.

This basic implementation is provided as a replacement of the original
algorithm. It can easily be replaced with an hardware architecture
optimized version. This version offers comparable performances in our
testings. Here are some measurements of the modpow execution time on
two different boards with both the original and the new algorithm.

On a Raptor Lake brya0 board:

| modpow() function call during boot | original | Algorithm 2 |
|------------------------------------+----------+-------------|
| coreboot/verstage 1                |    5.642 |       5.818 |
| coreboot/verstage 2                |    1.560 |       1.426 |
| coreboot/verstage 3                |    1.548 |       1.414 |
| depthcharge 1                      |    0.693 |       0.708 |
| depthcharge 2                      |    0.172 |       0.180 |
| depthcharge 3                      |    0.223 |       0.217 |
|------------------------------------+----------+-------------|
| Total (ms)                         |    9.838 |       9.763 |

On a Meteor Lake rex0 board:

| modpow() function call during boot | original | Algorithm 2 |
|------------------------------------+----------+-------------|
| coreboot/verstage  1               |    6.531 |       6.588 |
| coreboot/verstage  2               |    1.854 |       1.728 |
| coreboot/verstage  3               |    1.841 |       1.771 |
| depthcharge  1                     |    0.547 |       0.576 |
| depthcharge  2                     |    0.152 |       0.152 |
| depthcharge  3                     |    0.164 |       0.165 |
|------------------------------------+----------+-------------|
| Total (ms)                         |   10.925 |      10.979 |

Change-Id: I8739f94c0f6423f6e848bcae10b56e8be8e352cd
Co-authored-by: Jeremy Compostella <jeremy.compostella@intel.com>