updater: update: Support multiple Servos without --servo_port
When multiple Servos are connected to the host, even if there is only
one servod running, the flashrom still needs the serial name for the
raiden_debug_spi programmer to choose the right Servo.

As a result, we should always try to get the Servo serial even if no
--servo_port (or SERVOD_PORT, SERVOD_NAME) was specified.

Also fixed a bug from CL:3367518 that when SERVOD_NAME (or SERVOD_PORT)
is empty, some debug messages may try to print a NULL pointer as string
(%s).

BUG=b:308533993
TEST=make; run unit tests
BRANCH=None

Change-Id: Ibe924c27ae6706f535d3911934bab01e49436938
Signed-off-by: Hung-Te Lin <hungte@chromium.org>