futility: updater: Add optional serial number argument to --ccd
Right now if we use `futility update --ccd`, there is no way to specify
which DUT we want to update. This is rather confusing and inconvenient
for developers which have many DUTs connected.

Add an optional argument to `--ccd` so that the user can write
`futility update --ccd=<CCD SERIAL NUMBER>` to specify which DUT to
access. This applies to the `read`, `update`, `flash`, and `gbb`
commands.

This compliments CL:4995584, which makes `--servo` always add the
serial number.

BUG=b:308533993
TEST=emerge vboot_reference; futility flash --flash_info --ccd=xxx
     where xxx targeted two different DUTs
BRANCH=none

Change-Id: Icdba37aed1d709e5e7797bbe1577f5b1f7351c12
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>