sign_uefi: Stop checking if the private key exists
In order to support signing via pkcs11, the private key may be a pkcs11
URI rather than a local file path, so don't fail signing if the file
doesn't exist.

Signing will still fail at a later point if any key files are missing,
this early check is just a convenience to stop the script quickly if
something is missing.

BRANCH=None
BUG=b:286234394
TEST=make runtests
TEST=run sign_official_build.sh on a local build

Change-Id: Ibf398a60ffa0fd30da3ff7cdf8a8a7efce7524f0
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>