futility/cmd_show: Fix output for firmware body signature
The condition 'if (state || (sign_key && good_sig))' near the end of
the show_fw_preamble_buf() is about keyblock signature. However the
output printed in that if block is about body signature inside the
preamble. Since the information about the keyblock signature is already
printed in show_keyblock(), we don't need to print anything at the end
of the function. Fix the problem and add a test case for this.

Also fix the prefix of a parseable output message "data_key::invalid".

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: I9f0b2135b422933e6aa710fcdf6d816c7fad0936
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>