2lib: Expand vb2_kernel_params for GKI purposes
Introduce four new parameters to vb2_kernel_params API:
1. vendor_boot_offset
2. init_boot_offset
3. init_boot_size
4. vboot_cmdline_offset

First two parameters are offsets in the kernel_buffer
to the loaded {vendor/init}_boot partitions, respectively.

Regarding items 3 and 4 - with enabling AVB in vboot it may be necessary
to convey cmdline updates to the bootloader. Put this data in kernel
buffer and add a field with offset of this region to the struct.
Furthermore, add init_boot_size which will be used later on for
verification if regions don't overlap in memory layout.

This CL is based on crrev.com/c/5650814

BUG=none
TEST=With consecutive CLs applied build FW and boot android
BRANCH=main

Change-Id: I5de37f8ab3f2b6b861d01c335111944400876acb
Signed-off-by: Konrad Adamczyk <konrada@google.com>
Signed-off-by: Jan Dabros <dabros@google.com>
Signed-off-by: Grzegorz Bernacki <bernacki@google.com>