futility/cmd_show: Make 'show' return 0 for invalid kernel body
As we already have the more strict 'futility verify' command, which
returns 0 only if everything in the input file is valid (including hash,
signature), change 'futility show' to return 0 for kernel body
verification failure.

The 'show' subcommand will fail only if it cannot access all the
expected parts of the input file. For example, when the keyblock header
contains some offset exceeding the input file size, we won't be able to
extract corresponding part (such as signature). In that case, both
'show' and 'verify' subcommands will fail.

BUG=b:290287265
TEST=make runfutiltests
BRANCH=none

Change-Id: I22e5f64597376e2daead4029b90f37f8b6ac6561
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>