tests/futility/test_show_and_verify: Add test case for invalid data key
Add a test case for an invalid firmware vblock (fw_pre) with an invalid
data key. Modify the test script to also check the stdout even if
'futility show' is expected to fail.

Note that '--type fw_pre' needs to be passed for the futility command to
fail. Otherwise the invalid vblock file will be recognized as a keyblock
file.

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: Ie0cab87f6cf3bc150c20eda7785d6663b751a6ee
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>