Update the public key data structure
The new algorithm we have implemented relies on the modulus positive
inverse instead of the modulus negative inverse. This commits
implements the data structure required changes.

This commit should theoratically be part of the previous
one ("firmware/2lib: Montgomery mult. using modulus positive inverse")
as this change is required for the previous commit to build. At the
moment, it is kept serparate to keep the previous simple but they can
be squashed in later.

Change-Id: I4f6300bd9b1a6477134eb7b91eaf4f1de45c96f0
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>