futility/cmd_show: Make preamble parseable output consistent
The parseable output is inconsistent for firmware preambles and kernel
preambles. For firmware, "fw_pre::preamble::valid" will be shown, but
not "fw_pre::preamble::signature::valid". For kernel, it'll be the
opposite.

To make the output consistent (and also consistent with the keyblock
parseable output), print both (although right now their values will
always be the same).

Also fix a typo "kernel_preamble", which should be just "preamble".

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: I4d8e0e21f6d1624b35dacabaad8998db2ee30994
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>