Makefile: Fix firmware build for FIRMWARE_ARCH arm64 and friends
Makefile maintains a list of expected values of FIRMWARE_ARCH, but
"arm64" isn't one of them. As a result, since oak we've been building
the firmware lib with "-DCHROMEOS_ENVIRONMENT ${COMMON_FLAGS}".

For simplicity, add more ARM archs to the expected list and map all of
them to "arm".

BUG=none
TEST=emerge-geralt vboot_reference
BRANCH=none

Change-Id: Idbdf311530831fb11769223a0ad765b4cf1cb1d5
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>