Makefile: Always link libdl
It seems that according to the POSIX standard, dlopen() is in a separate
library called libdl. On newer versions of glibc, the functionality is
already integrated in the main libc and libdl is an empty compatibility
stub, that's why we didn't notice that we forgot to add this missing
dependency. But on older versions the libraries are actually separate,
so in order to make vboot build correctly for those we need to add this.

BRANCH=none
BUG=b:294369070
TEST=none

Change-Id: I0adbf80e8b4a96bc06d5072aefe9d4e637d692cd
Signed-off-by: Julius Werner <jwerner@chromium.org>