Makefile: Always link libdl
It seems that according to the POSIX standard, dlopen() is in a separate
library called libdl. On newer versions of glibc, the functionality is
already integrated in the main libc and libdl is an empty compatibility
stub, that's why we didn't notice that we forgot to add this missing
dependency. But on older versions the libraries are actually separate,
so in order to make vboot build correctly for those we need to add this.

BRANCH=none
BUG=b:294369070
TEST=none

(cherry picked from commit ba7fd22d3dd659b3ab42ad66f22169a6ce0f4dfe)

Change-Id: I0adbf80e8b4a96bc06d5072aefe9d4e637d692cd
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5062277
Original-Reviewed-by: uwe taz <uwetaz@gmail.com>
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
GitOrigin-RevId: ba7fd22d3dd659b3ab42ad66f22169a6ce0f4dfe
Cr-Build-Id: 8739997580243611025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739997580243611025
Copybot-Job-Name: android-main-fw-nissa-vboot_reference-copybot-downstream