image_signing: Add support for signing Flexor kernel image
This change adds support for signing a flexor vmlinuz
file using UEFI keys (residing in ${key_dir}/uefi).

This is done if a Flexor vmlinuz file is found on the rootfs
partition of a reven image. The flexor vmlinuz file is then
not signed in place on the rootfs, but instead *moved* out
to a flexor_vmlinuz.signed file and then signed there.

The flexor_vmlinuz.signed file will end up right next to the
${OUTPUT_IMAGE} location.

BRANCH=none
BUG=b:316321064, b:316859627
TEST=make runtests

Change-Id: If1f5f1526befb25168aad6bab14d15005abe0d21
Signed-off-by: Bastian Kersting <bkersting@google.com>