futility: update: Don't fail if --unlock_me is used on a locked device
CL:4951472 changed the updater to fail if --unlock_me is used on a
locked device. This was to fix the case where applying --unlock_me on
a locked device makes it appear as if the descriptor has changed, so we
switch to an RW-only update unconditionally.

However, this causes problems in factory, because the code which passes
--unlock_me is used for both factory (unlocked descriptor) and RMA
(locked descriptor). To workaround this, the factory code was changed to
only pass --unlock_me if the descriptor is not already locked
(CL:4976237, CL:5003930). But it would be better to handle this inside
futility, to reduce complexity in factory code.

Change the updater to only apply --unlock_me if the descriptor on flash
is not already locked. This is basically equivalent to the current
behaviour, since factory only passes --unlock_me if the descriptor is
unlocked anyway.

BRANCH=None
BUG=b:270275115
TEST=make runfutiltests

Change-Id: I177390d2c4c9ac62787eee8a9f70bfcc177bcf6c
Signed-off-by: Reka Norman <rekanorman@google.com>