futility/cmd_show: Add "::valid" summary to vblock parseable output
If we want to know whether VBLOCK_A (paired with FW_MAIN_A) within an AP
firmware image is fully valid, right now there is no simple command to
check it. The 'futility verify' command succeeds only if everything in
the firmware image is valid. The 'futility verify -P' also doesn't
contain a summary line for the vblocks. For now we'll need to grep 3
lines of output

 bios::VBLOCK_A::keyblock::signature::valid
 bios::VBLOCK_A::preamble::signature::valid
 bios::VBLOCK_A::body::signature::valid

to be sure about the validity of VBLOCK_A.

To make scripts such as dev_debug_vboot easier to parse the output, add
a line "bios::VBLOCK_A::valid" to firmware vblock parseable output, as
well as a line "kernel::valid" to kernel vblock parseable output.

BUG=b:305659154
TEST=make runfutiltests
BRANCH=none

Change-Id: I574e0a8348ce33781a5d229fa7737da6c37bfa71
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>