avb: get preloaded partitions
In order to speed up boot process and limit number of memcopy it is
beneficial to make use of get_preloaded* callback. Instead of IO
operations, one just need to provide pointer to the memory block which
is already there.

In order to store ramdisks content properly in the buffer provided to
bootloader, load both vendor_boot and init_boot partition simultanously.

BUG=b:309426555
TEST=With consecutive CLs applied build FW and boot android
BRANCH=main

Change-Id: I056d2976dffdfd90ea804144bf06144dde411536
Signed-off-by: Konrad Adamczyk <konrada@google.com>
Signed-off-by: Jan Dabros <dabros@google.com>
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Signed-off-by: Grzegorz Bernacki <bernacki@google.com>