avb: get preloaded partitions
In order to speed up boot process and limit number of memcopy it is
beneficial to make use of get_preloaded* callback. Instead of IO
operations, one just need to provide pointer to the memory block which
is already there.

In order to store ramdisks content properly in the buffer provided to
bootloader, load both vendor_boot and init_boot partition simultanously.

BUG=b:357901475
BRANCH=None
TEST=With consecutive CLs applied build FW and boot android

Change-Id: I49abdf26a8693f5a0dacdc405b4228b41e3d8110
Signed-off-by: Konrad Adamczyk <konrada@google.com>
Signed-off-by: Jan Dabros <dabros@google.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>