tests: Rename ENABLE_HWCRYPTO_RSA_TESTS to TEST_ENABLE_HWCRYPTO_RSA
This patch just renames a Makefile variable to a slightly different
name. Normally, if we were using well-designed tooling, this should have
no effect.

Unfortunately we don't, and the platform2_test.py tool that is used by
the ebuild to wrap our tests has a list of magic variable names that it
allows Makefiles to export to the underlying tests. If you're using a
name not on that list, your tests will work right when run locally but
not when run by the ebuild. ENABLE_HWCRYPTO_RSA_TESTS is not one of the
magic names, so we need to change it to a name that is.

BRANCH=none
BUG=none
TEST=cros_run_unit_tests --protect vboot_reference [--board <board>]

Change-Id: I812ead8f2360c6632ebfc669a8d38958a870d76d
Signed-off-by: Julius Werner <jwerner@chromium.org>