host/lib/cbfstool: Add cbfstool_get_config_string()
In the 'config' file in CBFS, values for Kconfig strings are always
enclosed with double quotes, as shown below:

 CONFIG_MAINBOARD_DIR="google/brya"

Add a new function cbfstool_get_config_string(), which is almost the
same as cbfstool_get_config_value(), except that the double quotes are
stripped. An error will be returned if the config value isn't enclosed
with double quotes.

BUG=none
TEST=cq
BRANCH=none

Change-Id: I62bfa4574d534407a4d746c384becbefa9381405
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>