sign_uefi: Handle case where the crdyshim key does not exist
Older reven keysets do not have the crdyshim key. That is fine: the
corresponding builds do not have crdyshim or crdyboot, so the key is not
needed.

Move the check for the key out of the early `check_keys`, and instead
check it only when signing crdyboot.

Also add a test for signing without the crdyshim key.

BRANCH=None
BUG=b:331773545
TEST=make runtests
TEST=Run sign_official_build.sh several times:
TEST=Success: crdyshim exist, build has crdy* bootloaders
TEST=Success: crdyshim keys removed, build has no crdy* bootloaders
TEST=Expected failure: crdyshim keys removed, build has crdy* bootloaders

Change-Id: I0912919edb0c1a7750c4434fcb00d99def79e771
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>