futility/dump_fmap: Exit with error if specified section is not found
The 'dump_fmap' command supports specifying a list of FMAP sections.
However, when some of the specified sections are not found in the image,
the command still exits with 0.

Fix it and also print out error messages for invalid section names.

Add a test case to test_dump_fmap.sh. Also remove an existing test case
for 'futility show', which doesn't belong to that file (and the case is
already covered in test_show_and_verify.sh).

BUG=none
TEST=cq
BRANCH=none

Change-Id: I40fb967adb6b67316008b63e97c103c4f451412f
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>