NISSA-ONLY: HACK: Record extra debugging info for vboot corruption
This patch adds a function that will check for secdata_kernel
corruption, and record several hints about the system state in the
recovery subcode field when it finds some. This is intended to help us
track down the source of secdata_kernel corruption that was observed on
Uldren and Pujjoga.

BRANCH=nissa
BUG=b:331989856
TEST=Will trybot for Nissa.

(cherry picked from commit 95727d53ea499541836a831a0a1cc21072373d41)

Change-Id: Ibbeef35a4e3c098b9afceb2530ed33415b3c05a2
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5499389
Original-Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5731257
Original-Reviewed-by: Subrata Banik <subratabanik@chromium.org>
GitOrigin-RevId: 95727d53ea499541836a831a0a1cc21072373d41