scripts/image_signing/swap_ec_rw: Always add ecrw.* as raw CBFS file
There is no need to read the CBFS file type of the existing "ecrw" and
use the same file type when adding the new "ecrw" to CBFS, because the
file type should always be "raw". The same argument applies to
"ecrw.hash" and "ecrw.version" as well. Therefore, always pass "-t raw"
for "ecrw.*".

BUG=b:325886850
TEST=make runtestscripts
BRANCH=none

Change-Id: Ie4fe80327f5f969bec10f0e3a3c74218d77edde8
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>