futility: Support ecrw version for --manifest
With go/swap-ecrw implemented, the "ecrw" version within the AP firmware
image may not come from the same build of the AP image. To allow easier
tracking of the "ecrw" version, add support for showing the ecrw version
for the 'futility update --manifest' command. An example output of
'futility update --manifest -i image-ciri.bin' is:

 {
   "default": {
     "host": {
       "versions": {
         "ro": "Google_Ciri.15878.0.0",
         "rw": "Google_Ciri.15878.0.0",
         "ecrw_a": "ciri-15699.0.0",
         "ecrw_b": "ciri-15699.0.0"
       },
       "keys": { ... },
       "image": "image-ciri.bin"
     }
   }
 }

Unlike the AP RW version, the ecrw versions in FW_MAIN_A and FW_MAIN_B
are separated in the output. Also, the output json format is slightly
changed.

BUG=b:340132454
TEST=futility update --manifest -i image-ciri.bin   # w/ ecrw.version
TEST=futility update --manifest -i image-hatch.bin  # w/o ecrw.version
BRANCH=none

Change-Id: I5e5591e0700a2b4c4b3cf6b82c552823f56fd1e4
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>