futility: Warn about inconsistent RW_FWID_A and RW_FWID_B versions
For simplicity, only the RW_FWID_A version is shown in the
'futility update --manifest' output. That avoids unnecessary
complication for futility users. However, that also makes it impossible
for the caller to know the exact versions of RW A and B.

Add a warning about inconsistent RW_FWID_A and RW_FWID_B versions.

BUG=none
TEST=make runtests -j
BRANCH=none

Change-Id: I03cb2ddb6d5d77cb8a3b51d0434f810f7bbb59b1
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>