extract_vmlinuz.c: Fix the bounds check on vmlinuz_header_{offset,size}
The check on vmlinuz_header_offset and vmlinuz_header_size is obviously
wrong:

	if (!vmlinuz_header_size ||
	    kpart_data + vmlinuz_header_offset + vmlinuz_header_size >
	    kpart_data) {
		return 1;
	}

`kpart_data + some_unsigned_values` can obviously never be `> kpart_data`,
unless something has overflowed!  And `vmlinuz_header_offset` hasn't even
been set yet (besides being initialized to zero)!

GCC will deduce that if the check didn't cause the function to bail, then
vmlinuz_header_size (a uint32_t) must be "negative"; that is: in the range
[2GiB,4GiB).

On platforms where size_t is 32-bits, this is *especially* broken.
memcpy's size argument must be in the range [0,2GiB).  Because GCC has
proved that vmlinuz_header_size is higher than that, it will fail to
compile:

	host/lib/extract_vmlinuz.c:67:9: error: 'memcpy' specified bound between 2147483648 and 4294967295 exceeds maximum object size 2147483647 [-Werror=stringop-overflow=]

So, fix the check.

I can now say that what I suspect the original author meant to write would
be the following patch, if `vmlinuz_header_offset` were already set:

	-kpart_data + vmlinuz_header_offset + vmlinuz_header_size > kpart_data
	+now        + vmlinuz_header_offset + vmlinuz_header_size > kpart_size

This hypothesis is supported by `now` not getting incremented by
`kblob_size` the way it is for the keyblock and preamble sizes.

However, we can also see that even this "corrected" bounds check is
insufficient: it does not detect the vmlinuz_header overflowing into
kblob_data.

OK, so let's describe the fix:

Have a `*vmlinuz_header` pointer instead of a
`uint64_t vmlinuz_header_offset`, to be more similar to all the other
regions.  With this change, the correct check becomes a simple

      vmlinuz_header + vmlinuz_header_size > kblob_data

While we're at it, make some changes that could have helped avoid this in
the first place:

 - Add comments.
 - Calculate the vmlinuz_header offset right away, instead of waiting.
 - Go ahead and increment `now` by `kblob_size`, to increase regularity.

Change-Id: I5c03e49070b6dd2e04459566ef7dd129d27736e4