swap_ec_rw: Use EC_RW flashmap section to get EC firmware
The RW_FW flashmap section is identical to EC_RW, but is for Zephyr
only (and is mainly an internal section that's necessary to get binman
to do the right thing).  EC_RW exists for both legacy EC and
Zephyr-based EC, and should get us the right thing.

BUG=b:352556657
BRANCH=none
TEST=swap_ec_rw with legacy EC based image

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I94a0550b79880757e9616ba13ea41ed90cc01625