futility/load_fmap: Erase remaining bytes if file smaller than area
The `futility load_fmap` subcommand allows the passed file to be smaller
than the corresponding FMAP area size. That's convenient for some use
cases (such as EC's KEY_RO), because the caller doesn't need to have an
extra step of enlarging the file to match the area size.

However, the current implementation is to keep the remaining data in the
area unmodified, which is often not the intention of the caller.
Therefore, change the behavior by erasing the remaining bytes to 0xff.

BUG=none
TEST=make runfutiltests
BRANCH=none

(cherry picked from commit 8365d546ce0698faffd22cf3cb44c0c854f74637)

Change-Id: Ib6575023801d995b7b25110bb411ec74d9a240f0
Original-Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5768173
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
GitOrigin-RevId: 8365d546ce0698faffd22cf3cb44c0c854f74637
Cr-Build-Id: 8737902702090947393
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737902702090947393
Copybot-Job-Name: android-main-vboot_reference-copybot-downstream