futility: updater: remove the unused quirk 'ec_partial_recovery'
The 'ec_partial_recovery' quirk was added to prevent resetting USB when
updating EC firmware. However that was never enabled because there were
other bugs in the flashrom.

Now we plan to move the EC updating to a dedicated command (ectool
updateimage), so it is time to remove this unused quirk.

BUG=b:218612817,b:152925812
TEST=make; build and run test
BRANCH=None

Change-Id: I2cee51b9e3bad1f89fa7e47ff4ef90e220bdd6a2
Signed-off-by: Hung-Te Lin <hungte@chromium.org>