avb: fix subtraction overflow in init boot offset calculation
This small patch swaps subtraction operants in init_boot_size. The later
operant always had a greater value than the left operant. This causes
the value to overflow and reach the values in neighborhood of ~4GB,
which is invalid.

This is a nit rather than fix, since the sole use of the value is added
back to init_boot_offset, which overflows back to a correct value.

BUG=b:359340876
TEST=Build FW and boot android
BRANCH=firmware-android-15949.B

Change-Id: I838675bed8dfb94dfdce4f9c1e147cc3128a04b6
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>